policy sync/policy-sync model/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reggenai"
version = "0.1.0"
description = "Autonomous Regulatory Sentinel - RAG pipeline for legal compliance"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    "python-dotenv",
    "pdfplumber",
    "python-docx",
    "beautifulsoup4",
    "lxml",
    "sentence-transformers",
    "faiss-cpu",
    "openai",
    "transformers",
    "torch",
    "langchain",
    "dateparser",
    "chromadb",
    "jsonschema",
    "sqlalchemy",
    "alembic",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black", "ruff"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["reggenai"]

[tool.setuptools.package-data]
reggenai = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
